Updated service & controller files
Error handling

+++ src/todos/todos.service.ts
@@ -1,65 +1,69 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
-import { Todo } from './todo.entity';
-import { Repository } from 'typeorm';
-import { CreateTodoDto } from 'src/todo/dto/create-todo.dto';
+import { Injectable, NotFoundException } from "@nestjs/common";
+import { InjectRepository } from "@nestjs/typeorm";
+import { Todo } from "./todo.entity";
+import { Repository } from "typeorm";
+import { CreateTodoDto } from "src/todo/dto/create-todo.dto";
+import { UpdateTodoDto } from "src/todo/dto/update-todo.dto";
 //import { UpdateTodoDto } from 'src/todos/dtos/update-todo.dto';
 //import { CreateTodoDto } from './dtos/create-todo.dto';
 
 @Injectable()
 export class TodosService {
-  update //     Object.assign(todo, dto);
-    (arg0: number, dto: CreateTodoDto) {
-      throw new Error('Method not implemented.');
-  }
-  
   constructor(
-    @InjectRepository(Todo) private readonly todoRepository: Repository<Todo>,
+    @InjectRepository(Todo) private readonly todoRepository: Repository<Todo>
   ) {}
 
   async create(dto: CreateTodoDto) {
-    const todo = this.todoRepository.create(dto);
+    try {
+      const todo = this.todoRepository.create(dto);
 
-    return await this.todoRepository.save(todo);
+      let res = await this.todoRepository.save(todo);
+
+      return res;
+    } catch (error) {
+      throw new Error(error);
+    }
   }
 
   async findMany() {
-    return this.todoRepository.find( );
+    return this.todoRepository.find();
   }
 
-  async findOne(id: number){
-    return this.todoRepository.findOne({where:{id}});
+  async findOne(id: number) {
+    if (!id) {
+      throw new Error("Not found: id is empty. ");
+    }
+    try {
+      return this.todoRepository.findOne({ where: { id } });
+    } catch (error) {
+      console.debug(error);
+      return false;
+    }
   }
 
+  async update(id: number, dto: UpdateTodoDto) {
+    if (!id) {
+      throw new Error("update error: id is empty. ");
+    }
+    try {
+      dto.id = id;
 
-  //   async update(id: number, dto: CreateTodoDto) {
-  //     const todo = this.todoRepository.findOne({ where: { id } });
-
-  //     Object.assign(todo, dto);
-
-  //     await this.todoRepository.save(todo);
-
-  //     return todo;
-  //   }
-
-//   async update(id: number, dto: CreateTodoDto): Promise<Todo> {
-//     const updatedRes = this.todoRepository.findOne({ where: { id } });
-
-//     (await updatedRes).title = dto.title;
-
-//     return await this.todoRepository.save(updatedRes);
-//   }
-
-
-// async updateTodo(id:number, dto: UpdateTodoDto){
-//   const updateDto = dto.id;
-//   return this.todoRepository.update(updateDto, dto);
-// }
+      await this.todoRepository.save(dto);
+      return `Updated successfully!`;
+    } catch (error) {
+      console.debug(error);
+      return false;
+    }
+  }
 
   async remove(id: number): Promise<void> {
-    await this.todoRepository.delete(id);
-    
+    if (!id) {
+      throw new Error("Not found error: id is empty. ");
+    }
+    try {
+      await this.todoRepository.delete(id);
+    } catch (error) {
+      console.debug(error);
+    }
   }
 }
-
-
